importers/apkindex: do not unpublish all packages
This reverts cb0d588 (importers: import-apkindex: mark packages as
unpublished before importing an apkindex, 2021-04-22).

The goal of having packages unpublished is that fixes should only be
published as soon as the package is available on the mirrors, not when
the commit has been made.

Once a fix has been published, there should be no reason anymore to
unpublish it, resulting it being invislble.

The expected flow after this change is:

1. import-secfixes creates a new package_version if it does not exist
   yet, and `published` defaults to False.
2. import-apkindex will set `published` to True for each package in the
   index.

So this means once `publshed` is set to True, it will never be set to
`False` anymore.

One issue that still remains is that when a secfix is published for an
older version of the package, it will never be published, since that
version no longer exists in the apkindex.