package_version: add succeeded flag
The goal of the published flag is to hide fixed versions until the
apk file is actually available. It is also used to mark versions as no
longer available implying that issues with it have been resolved.

The problem is that this hides any vulnerability information on older
versions, including fixes.

By adding a separate flag, we can distinguish between the two
situations:

Once the apk is present, published will be set to true and will always
remain true. Once the package version is no longer in the apkindex,
succeeded is set to true.

Therefore, a vulnerable version is one where published is true, and
succeeded is false.

This requires a database migration:

```
ALTER TABLE package_version ADD COLUMN succeeded BOOLEAN DEFAULT 0 CHECK (succeeded IN (0, 1));
CREATE INDEX ix_package_version_succeeded ON package_version (succeeded);
```