vulnrich: check for version type
The affected versions has a type, which tells us how to interpret the
versions and comparissons. 'semver' and 'python' are the most usefull
ones. The other version types are either not relevant to us (eg. 'rpm'),
or not reliable (ie. 'custom').

We do keep the version in case the version type is not supported, but we
only do equal comparissons to prevent false positive matches.

See:
- https://github.com/CVEProject/cve-schema/blob/main/schema/docs/versions.md
- https://github.com/CVEProject/cve-schema/blob/main/schema/docs/CVE_Record_Format_bundled.json